pyproject.toml
[project]
name = "my-hello-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "hello-agents==0.1.1",
    "langchain>=1.2.15",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.6",
    "openai>=2.30.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily>=1.1.0",
]

[project.scripts]
hello_agents_llm = "my_hello_agent.hello_agents_llm:main"

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
